NOT COMPATIBLE BACKWARD!!!! CONNECTION CHANGED
git-svn-id: https://svn.code.sf.net/p/beebeep/code@30 127b15f6-8070-407b-945c-913c06b37495

+++ Connection.h
@@ -71,7 +71,7 @@ private:
   State m_state;
   bool m_isHelloMessageSent;
 
-  quint16 m_rawDataSize;
+  quint32 m_rawDataSize;
 };
